+++ src/Driver/Postgres.php
@@ -13,6 +13,7 @@ declare(strict_types=1);
  * @license     https://opensource.org/licenses/mit-license.php MIT License
  * @since       2.1.0
  */
+
 namespace DatabaseBackup\Driver;
 
 /**