+++ tests/TestCase/BackupTraitTest.php
@@ -1,5 +1,4 @@
 <?php
-/** @noinspection PhpUnhandledExceptionInspection */
 declare(strict_types=1);
 
 /**
@@ -13,38 +12,44 @@ declare(strict_types=1);
  * @link        https://github.com/mirko-pagliai/cakephp-database-backup
  * @license     https://opensource.org/licenses/mit-license.php MIT License
  */
+
 namespace DatabaseBackup\Test\TestCase;
 
+use App\BackupTraitAsClass;
 use Cake\Core\Configure;
 use Cake\Database\Connection;
 use Cake\Datasource\ConnectionManager;
 use Cake\Datasource\Exception\MissingDatasourceConfigException;
-use DatabaseBackup\BackupTrait;
 use DatabaseBackup\TestSuite\TestCase;
 
 /**
  * BackupTraitTest class
+ *
+ * @uses \DatabaseBackup\BackupTrait
  */
 class BackupTraitTest extends TestCase
 {
     /**
-     * @psalm-var trait-string<\DatabaseBackup\BackupTrait>
+     * @var \App\BackupTraitAsClass
      */
-    protected $Trait;
+    protected BackupTraitAsClass $Trait;
 
     /**
      * @inheritDoc
      */
-    public array $fixtures = ['core.Articles', 'core.Comments'];
+    public array $fixtures = [
+        'core.Articles',
+        'core.Comments',
+    ];
 
     /**
      * @inheritDoc
      */
-    public function setUp(): void
+    protected function setUp(): void
     {
         parent::setUp();
 
-        $this->Trait ??= $this->getMockForTrait(BackupTrait::class);
+        $this->Trait ??= new BackupTraitAsClass();
     }
 
     /**