+++ tests/TestCase/Driver/SqliteTest.php
@@ -12,20 +12,22 @@ declare(strict_types=1);
  * @link        https://github.com/mirko-pagliai/cakephp-database-backup
  * @license     https://opensource.org/licenses/mit-license.php MIT License
  */
+
 namespace DatabaseBackup\Test\TestCase\Driver;
 
 use DatabaseBackup\TestSuite\DriverTestCase;
 
 /**
  * SqliteTest class
+ *
+ * @uses \DatabaseBackup\Driver\Sqlite
  */
 class SqliteTest extends DriverTestCase
 {
     /**
-     * Called before every test method
-     * @return void
+     * @inheritDoc
      */
-    public function setUp(): void
+    protected function setUp(): void
     {
         parent::setUp();