+++ tests/TestCase/Utility/BackupImportTest.php
@@ -1,5 +1,4 @@
 <?php
-/** @noinspection PhpUnhandledExceptionInspection */
 declare(strict_types=1);
 
 /**
@@ -13,6 +12,7 @@ declare(strict_types=1);
  * @link        https://github.com/mirko-pagliai/cakephp-database-backup
  * @license     https://opensource.org/licenses/mit-license.php MIT License
  */
+
 namespace DatabaseBackup\Test\TestCase\Utility;
 
 use Cake\Event\EventList;
@@ -24,6 +24,8 @@ use Symfony\Component\Process\Process;
 
 /**
  * BackupImportTest class
+ *
+ * @uses \DatabaseBackup\Utility\BackupImport
  */
 class BackupImportTest extends TestCase
 {
@@ -33,10 +35,10 @@ class BackupImportTest extends TestCase
     protected BackupImport $BackupImport;
 
     /**
-     * Called before every test method
-     * @return void
+     * {@inheritDoc}
+     * @throws \ReflectionException
      */
-    public function setUp(): void
+    protected function setUp(): void
     {
         parent::setUp();
 
@@ -47,6 +49,7 @@ class BackupImportTest extends TestCase
     /**
      * Test for `filename()` method. This tests also `$compression` property
      * @test
+     * @throws \ReflectionException
      * @uses \DatabaseBackup\Utility\BackupImport::filename()
      */
     public function testFilename(): void
@@ -90,6 +93,7 @@ class BackupImportTest extends TestCase
 
     /**
      * @test
+     * @throws \ReflectionException
      * @uses \DatabaseBackup\Utility\BackupImport::import()
      */
     public function testImport(): void
@@ -110,6 +114,8 @@ class BackupImportTest extends TestCase
     /**
      * Test for `import()` method. Export is stopped by the `Backup.beforeImport` event (implemented by driver)
      * @test
+     * @throws \PHPUnit\Framework\MockObject\Exception
+     * @throws \ReflectionException
      * @uses \DatabaseBackup\Utility\BackupImport::import()
      */
     public function testImportStoppedByBeforeExport(): void
@@ -125,6 +131,8 @@ class BackupImportTest extends TestCase
     /**
      * Test for `import()` method, on failure (error for `Process`)
      * @test
+     * @throws \PHPUnit\Framework\MockObject\Exception
+     * @throws \ReflectionException
      * @uses \DatabaseBackup\Utility\BackupImport::import()
      */
     public function testImportOnFailure(): void
@@ -142,6 +150,8 @@ class BackupImportTest extends TestCase
      * @see https://symfony.com/doc/current/components/process.html#process-timeout
      * @test
      * @uses \DatabaseBackup\Utility\BackupImport::import()
+     * @throws \PHPUnit\Framework\MockObject\Exception
+     * @throws \ReflectionException
      */
     public function testImportExceedingTimeout(): void
     {