+++ tests/test_app/TestApp/Application.php
@@ -1,5 +1,4 @@
 <?php
-/** @noinspection PhpMissingDocCommentInspection */
 declare(strict_types=1);
 
 namespace App;